Moved HomeChurchItem, HomeChurchLocationSelect to separate files created HomeChurchMapScreen UI
