+++ package.json
@@ -58,6 +58,7 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
     "react-native-expo-cached-image": "^1.3.1",
     "react-native-gesture-handler": "~1.8.0",
+    "react-native-maps": "0.27.1",
     "react-native-open-maps": "^0.3.5",
     "react-native-reanimated": "~1.13.0",
     "react-native-safe-area-context": "3.1.9",
